Changed mistaken uint8_t to uint16_t

+++ libs/libcmdtlm/commands.hpp
@@ -6,7 +6,7 @@
 class Commands {
 public:
   virtual void control(const ControlPacketElement &e) {}
-  virtual void lwirFrame(const uint8_t frame[60][80]) {}
+  virtual void lwirFrame(const uint16_t frame[60][80]) {}
 };
 
 #endif
